added conditional rendering and fixed a bug with the variable labels that is seen if you change scene collections after displaying the variable definition dialog
