+++ src/obs-text.hpp
@@ -1,2 +1,149 @@
-bool InitOBSText();
-void FreeOBSText();
+#include <util/platform.h>
+#include <combaseapi.h>
+#include <gdiplus.h>
+#include <string>
+#include <memory>
+#include <plugin-support.h>
+
+using namespace std;
+using namespace Gdiplus;
+
+enum class Align {
+	Left,
+	Center,
+	Right,
+};
+
+enum class VAlign {
+	Top,
+	Center,
+	Bottom,
+};
+
+template<typename T, typename T2, BOOL WINAPI deleter(T2)> class GDIObj {
+	T obj = nullptr;
+
+	inline GDIObj &Replace(T obj_)
+	{
+		if (obj)
+			deleter(obj);
+		obj = obj_;
+		return *this;
+	}
+
+public:
+	inline GDIObj() {}
+	inline GDIObj(T obj_) : obj(obj_) {}
+	inline ~GDIObj() { deleter(obj); }
+
+	inline T operator=(T obj_)
+	{
+		Replace(obj_);
+		return obj;
+	}
+
+	inline operator T() const { return obj; }
+
+	inline bool operator==(T obj_) const { return obj == obj_; }
+	inline bool operator!=(T obj_) const { return obj != obj_; }
+};
+
+using HDCObj = GDIObj<HDC, HDC, DeleteDC>;
+using HFONTObj = GDIObj<HFONT, HGDIOBJ, DeleteObject>;
+using HBITMAPObj = GDIObj<HBITMAP, HGDIOBJ, DeleteObject>;
+
+#define S_TRANSFORM_NONE 0
+
+struct TextSource {
+	obs_source_t *source = nullptr;
+
+	gs_texture_t *tex = nullptr;
+	uint32_t cx = 0;
+	uint32_t cy = 0;
+
+	HDCObj hdc;
+	Graphics graphics;
+
+	HFONTObj hfont;
+	unique_ptr<Font> font;
+
+	bool read_from_file = false;
+	string file;
+	time_t file_timestamp = 0;
+	bool update_file = false;
+	float update_time_elapsed = 0.0f;
+
+	wstring text;
+	wstring text_to_render;
+	wstring face;
+	int face_size = 0;
+	uint32_t color = 0xFFFFFF;
+	uint32_t color2 = 0xFFFFFF;
+	float gradient_dir = 0;
+	uint32_t opacity = 100;
+	uint32_t opacity2 = 100;
+	uint32_t bk_color = 0;
+	uint32_t bk_opacity = 0;
+	Align align = Align::Left;
+	VAlign valign = VAlign::Top;
+	bool gradient = false;
+	bool bold = false;
+	bool italic = false;
+	bool underline = false;
+	bool strikeout = false;
+	bool antialiasing = true;
+	bool vertical = false;
+
+	bool use_outline = false;
+	float outline_size = 0.0f;
+	uint32_t outline_color = 0;
+	uint32_t outline_opacity = 100;
+
+	bool use_extents = false;
+	bool wrap = false;
+	uint32_t extents_cx = 0;
+	uint32_t extents_cy = 0;
+
+	int text_transform = S_TRANSFORM_NONE;
+
+	bool chatlog_mode = false;
+	int chatlog_lines = 6;
+
+	/* --------------------------- */
+
+	inline TextSource(obs_source_t *source_, obs_data_t *settings)
+		: source(source_),
+		  hdc(CreateCompatibleDC(nullptr)),
+		  graphics(hdc)
+	{
+		obs_source_update(source, settings);
+	}
+
+	inline ~TextSource()
+	{
+		if (tex) {
+			obs_enter_graphics();
+			gs_texture_destroy(tex);
+			obs_leave_graphics();
+		}
+	}
+
+	void UpdateTextToRender();
+	void UpdateFont();
+	void GetStringFormat(StringFormat &format);
+	void RemoveNewlinePadding(const StringFormat &format, RectF &box);
+	void CalculateTextSizes(const StringFormat &format, RectF &bounding_box,
+				SIZE &text_size);
+	void RenderOutlineText(Graphics &graphics, const GraphicsPath &path,
+			       const Brush &brush);
+	void RenderText();
+	void LoadFileText();
+	void TransformText();
+	void SetAntiAliasing(Graphics &graphics_bitmap);
+
+	const char *GetMainString(const char *str);
+
+	inline void Update(obs_data_t *settings);
+	inline void Tick(float seconds);
+	inline void Render();
+};
