+++ src/plugin-lifetime.hpp
@@ -1,0 +1,2 @@
+bool InitOBSText();
+void FreeOBSText();
