+++ src/variables.cpp
@@ -1,3 +1,4 @@
+#include <util/util.hpp>
 #include "variables.hpp"
 
 using namespace std;
@@ -14,7 +15,8 @@ VariablesAndValues *VariablesAndValues::getInstance()
 
 void VariablesAndValues::putVariable(const QString &variable)
 {
-
+	blog(LOG_INFO, "VariablesAndValues::putVariable %s",
+	     variable.toStdString().c_str());
 	if (variablesAndValues.find(variable) == variablesAndValues.end()) {
 		variablesAndValues[variable] = QString("");
 		variables.insert(variable);
@@ -28,6 +30,8 @@ bool VariablesAndValues::contains(const QString &variable)
 
 void VariablesAndValues::putValue(const QString &variable, const QString &value)
 {
+	blog(LOG_INFO, "VariablesAndValues::putValue %s %s",
+	     variable.toStdString().c_str(), value.toStdString().c_str());
 	variablesAndValues[variable] = value;
 	variables.insert(variable);
 }
@@ -44,6 +48,8 @@ const QString &VariablesAndValues::getValue(const QString &variable)
 
 void VariablesAndValues::clear()
 {
+	blog(LOG_INFO, "VariablesAndValues::clear");
+
 	variablesAndValues.clear();
 	variables.clear();
 }
